Give D_8009B0D8 one type, and say what its two remaining arms are for
The declaration chain for this symbol offered four arms at three widths --
u8 in .data, u16, volatile s32, s32 -- and two translation units carried
private asm-aliased spellings beside it. That reads as a type nobody could
pin down. Measured, it is one four-byte frame step, and the widths came from
the use sites rather than from any disagreement.

The census, over the full function listings of the executable in the other
matching decompilation, and over this repository's candidate and overlay
targets for the offset check: eighteen accesses across twelve functions,
every one at offset zero. Thirteen are word, three
halfword, two byte; by function that is nine, two and two, with
func_80012DB4 in two of the groups. No D_8009B0D8+N appears in any resident
or overlay target, and no data table takes its address. c_symbols.ld puts
the next named symbol at D_8009B0E8, sixteen bytes on, and nothing reads
into that gap. The value is written once, as D_8009B0D8 = D_8009B0C1 + 1 in
graphics_frame.c, and every other consumer scales a motion delta by it --
* 2, * 8, * 16, << 3 -- which is a per-frame step count.

Three widths out of one declaration is what gcc 2.8.1 does on its own. A
five-line probe through the project compiler:

    extern int g;
    int wide(void)    { return g; }                  ->  lw   $2,g
    int lowbyte(void) { return (unsigned char)g; }   ->  lbu  $2,g
    int lowhalf(void) { return (unsigned short)g; }  ->  lhu  $2,g

The narrowing is chosen at the use. The other matching decompilation of this
executable is the control that this holds for this symbol and not only in
the abstract: its func_80012DB4 declares a plain s32 scalar, writes
D_8009B0D8 = D_8009B0C1 + 1 and reads (u8)D_8009B0D8, and its listing
carries both sw and lbu against %gp_rel for this address.

So the widths collapse. Each step below was verified by its own clean build
before the next was applied:

- func_80041D60.c drops D_8009B0D8_IS_HALFWORD and casts at the use,
  (u16)arg0->field_5A - (u16)D_8009B0D8.
- The .data arm becomes s32, and input_pads.c casts at the use,
  gInput_abRepeatTimers[i] + (u8)D_8009B0D8.
- display_effect_update_callbacks.c drops
  extern u16 D_8009B0D8_halfword asm("D_8009B0D8") and casts at its two
  sites.
- display_object_fade_callbacks.c drops
  extern volatile s32 D_8009B0D8_volatile asm("D_8009B0D8") and takes
  D_8009B0D8_IS_VOLATILE for the whole unit; its three plain reads build the
  same under it. Rewriting those two sites as *(volatile s32 *)&D_8009B0D8
  instead does not work -- .text grows until .initialized_data overlaps it
  and the executable will not link -- so the unit-wide guard is the spelling,
  not a cast.
- The u16 arm leaves the header, having no definer left, and func_8003CDF8.c
  and func_8003CE48.c defined D_8009B0D8_IN_DATA without referencing the
  symbol at all, so those two defines go too.

What remains is two arms and a default, all s32: the plain one, the same in
.data, and the volatile one. Section placement and volatility are addressing
and ordering properties that no cast at a use site can express, which is
what those arms are for, and the header now says so. No asm alias for this
symbol is left anywhere in the tree.

For scale, since this symbol was raised as an instance of a general problem:
of 828 symbols declared in headers, 73 carry a multi-arm chain. The same
width test that flagged this one now flags exactly one other, D_801D0000 --
and that one is the save-data workspace, declared s16[] or u8[] by a guard
whose comment gives the reason, so it is an element type on a buffer rather
than a scalar read two ways. The other 71 split on .data (32), volatile
(23), array-versus-scalar (9) and other (8): addressing and ordering, not
type.

main_menu/func_80180390 references the symbol, so its recorded canonical
contract moves with the declaration; the hash is the one
make candidate-contract-hashes prints. Its candidate_build_sha256 and
target_bytes_sha256 are unchanged and make check-candidate-builds reports OK
for all 171.

make clean match, check-notes, check-metadata, basic-types,
external-attempts and check-candidate-builds all pass, at sha256
84a54ed74f3d0edd6d81380839f7e4ef5bfb21ecea18be9a062bd6bfa5a45c88.

+++ src/game/input_pads.c
@@ -106,7 +106,7 @@ void Input_UpdatePads(void)
             if (new_bits & INPUT_PENDING_HIGH_BIT) {
                 repeat |= 1;
             }
-            value = gInput_abRepeatTimers[i] + D_8009B0D8;
+            value = gInput_abRepeatTimers[i] + (u8)D_8009B0D8;
             gInput_abRepeatTimers[i] = value;
             if (value >= gInput_bRepeatDelay) {
                 gInput_abRepeatTimers[i] = gInput_bRepeatInterval;
